Remove tzinfo-data dependency

+++ db/migrate/20181027042510_remove_useless_models.rb
@@ -1,6 +1,7 @@
 class RemoveUselessModels < ActiveRecord::Migration[5.1]
-  def change
-		drop_table :scrapers
-		drop_table :site_scrapers
-  end
+	def change
+		drop_table :scrapers if table_exists? :scrapers
+		drop_table :site_scrapers if table_exists? :site_scrapers
+	end
 end
+
